Move aesthetics into shared folder, this makes the server config more complex. Possibly reconsider.
