+++ src/index.ts
@@ -35,6 +35,7 @@ import {
 import {
   serveReport,
   serveStaticResource,
+  serveStaticSharedResource,
   serveViewJs
 } from './backend/dev-server/server.js';
 import { DatabaseWithPool } from './backend/db/database-with-pool.js';
@@ -134,6 +135,7 @@ router.post(
 
 if (DEV) {
   router.get(`${siteConfig.staticUrl}/:filename*`, serveStaticResource);
+  router.get(`/shared/:filename*`, serveStaticSharedResource);
   router.get(`/src/frontend/:filename*`, serveViewJs);
   router.get(
     `${siteConfig.reportsUrl}/:change/figure-html/:filename`,
